chore: design doc for aurgy backend (#14)
this is a preliminary design doc for the backend services. it is **not** comprehensive but should give a general idea of how to attack the backend. It is missing some components, such as invite links, lobby management, etc. I think these are problems we can add later and should be discussed with the frontend team.

I also didn't do any caching/optimizations in this doc, we can add those later!